wip: fixed ad block, now issue with map getbounds
